added shadcn dialog component for reuse instead of default radixui
